Move scoreboard cache to the world cache
