+++ connector/src/main/java/org/geysermc/connector/network/translators/java/scoreboard/JavaDisplayScoreboardTranslator.java
@@ -36,7 +36,7 @@ public class JavaDisplayScoreboardTranslator extends PacketTranslator<ServerDisp
 
     @Override
     public void translate(ServerDisplayScoreboardPacket packet, GeyserSession session) {
-        session.getScoreboardCache().getScoreboard().registerNewObjective(
+        session.getWorldCache().getScoreboard().registerNewObjective(
                 packet.getName(), packet.getPosition()
         );
     }
